+++ frontend/src/Welcome/Navigation.js
@@ -1,0 +1,40 @@
+import React from 'react';
+import { Nav, MenuItem, Navbar, NavItem, NavDropdown} from 'react-bootstrap';
+
+function Navi() {
+    return <Navbar className="sticky-nav" collapseOnSelect>
+        <Navbar.Header>
+            <Navbar.Brand>
+                <a href="#brand">Home</a>
+            </Navbar.Brand>
+            <Navbar.Toggle />
+        </Navbar.Header>
+        <Navbar.Collapse>
+            <Nav>
+                <NavItem eventKey={1} href="#">
+                    <strong>Link</strong>
+                </NavItem>
+                <NavItem eventKey={2} href="#">
+                    Link
+          </NavItem>
+                <NavDropdown eventKey={3} title="Dropdown" id="basic-nav-dropdown">
+                    <MenuItem eventKey={3.1}>Action</MenuItem>
+                    <MenuItem eventKey={3.2}>Another action</MenuItem>
+                    <MenuItem eventKey={3.3}>Something else here</MenuItem>
+                    <MenuItem divider />
+                    <MenuItem eventKey={3.3}>Separated link</MenuItem>
+                </NavDropdown>
+            </Nav>
+            <Nav pullRight>
+                <NavItem eventKey={1} href="#">
+                    Link Right
+          </NavItem>
+                <NavItem eventKey={2} href="#">
+                    Link Right
+          </NavItem>
+            </Nav>
+        </Navbar.Collapse>
+    </Navbar>;
+}
+
+export default Navi;
